+++ .travis.yml
@@ -2,7 +2,7 @@ language: csharp
 dotnet: 2.1.4
 script:
   - dotnet restore
-  - dotnet build
+  - dotnet build --configuration Release
 deploy:
   skip_cleanup: true
   provider: script
