+++ SqlHelper.sln
@@ -1,32 +1,22 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
-VisualStudioVersion = 15.0.27130.2036
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SqlWrapperLite", "SqlHelper\SqlWrapperLite.csproj", "{69D21E17-8955-4F81-8400-99D9FA48BBA6}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{60C46133-81C2-433F-9067-DD56659472B0}"
-	ProjectSection(SolutionItems) = preProject
-		.travis.yml = .travis.yml
-		deploy.sh = deploy.sh
-		README.md = README.md
-	EndProjectSection
+VisualStudioVersion = 15.0.26124.0
+MinimumVisualStudioVersion = 15.0.26124.0
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SqlWrapperLite", "SqlWrapperLite\SqlWrapperLite.csproj", "{80F3C19B-DE30-43D4-8E49-DB8DB18B92B2}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{69D21E17-8955-4F81-8400-99D9FA48BBA6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{69D21E17-8955-4F81-8400-99D9FA48BBA6}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{69D21E17-8955-4F81-8400-99D9FA48BBA6}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{69D21E17-8955-4F81-8400-99D9FA48BBA6}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {525BA3C0-942D-4703-8CDB-AF1E0A12E849}
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{80F3C19B-DE30-43D4-8E49-DB8DB18B92B2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{80F3C19B-DE30-43D4-8E49-DB8DB18B92B2}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{80F3C19B-DE30-43D4-8E49-DB8DB18B92B2}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{80F3C19B-DE30-43D4-8E49-DB8DB18B92B2}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
